+++ packages/vue-ui/src/index.ts
@@ -1,0 +1,26 @@
+/**
+ * @arcships/vue-ui
+ *
+ * Vue 3 UI components for document-centric applications.
+ */
+
+// Signature pad
+export { default as SignaturePad } from "./components/SignaturePad.vue"
+
+// File upload
+export { default as FileUpload } from "./components/FileUpload.vue"
+
+// File thumbnail
+export { default as FileThumbnail } from "./components/FileThumbnail.vue"
+
+// Bounding box citations
+export { default as BoundingBoxCitations } from "./components/BoundingBoxCitations.vue"
+
+// Layout blocks
+export { default as LayoutBlocks } from "./components/LayoutBlocks.vue"
+
+// Base UI
+export { default as Spinner } from "./components/Spinner.vue"
+export { default as Tooltip } from "./components/Tooltip.vue"
+
+export type * from "./types"