+++ tests/consumer/template/src/main.ts
@@ -1,7 +1,7 @@
 import { createApp } from "vue";
 import "@arcships/vue-docx/style.css";
 import "@arcships/vue-xlsx/style.css";
-import "@arcships/vue-extend/style.css";
+import "@arcships/vue-pdf/style.css";
 import App from "./App.vue";
 import "./package-contract";